pyproject.toml
[project]
name = "memoire-stage"
version = "0.1.0"
description = ""
authors = [
    {name = "Geobatpo07",email = "lgeobatpo98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.2.6,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "sympy (>=1.14.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "geopandas (>=1.0.1,<2.0.0)",
    "cartopy (>=0.24.1,<0.25.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
